refactor: improve conversation file
