+++ app/(dashboard)/mailboxes/[mailbox_slug]/[category]/conversation/assignPopoverButton.tsx
@@ -12,7 +12,7 @@ import { useSession } from "@/components/useSession";
 import { getFullName } from "@/lib/auth/authUtils";
 import { cn } from "@/lib/utils";
 import { api } from "@/trpc/react";
-import { useAssignTicket } from "./useAssignTicket";
+import { useAssignTicket } from "./hooks/use-assign-ticket";
 
 export const AssignPopoverButton = ({
   initialAssignedToId,
