+++ app/(dashboard)/mailboxes/[mailbox_slug]/[category]/conversation/hooks/use-undone-email-store.ts
@@ -1,0 +1,18 @@
+import { create } from "zustand";
+import { devtools } from "zustand/middleware";
+import { DraftedEmail } from "@/app/types/global";
+
+export const useUndoneEmailStore = create<{
+  undoneEmail: DraftedEmail | undefined;
+  setUndoneEmail: (undoneEmail: DraftedEmail | undefined) => void;
+}>()(
+  devtools(
+    (set) => ({
+      undoneEmail: undefined,
+      setUndoneEmail: (undoneEmail) => set({ undoneEmail }),
+    }),
+    {
+      name: "undone-email-store",
+    },
+  ),
+);
